// src/model/claim.ts
export class Claim {
  id: number = 0;
  claimNumber: string = '';
  creationDate: Date = new Date();
  status: number = 0;
  statusName: string = '';
  name: string = '';
  identification: string = '';
  phone: string = '';
  dateOfIncident: Date = new Date();
  mount: string = '' ;
  userCreation: number = 0;
}